feat(packages): carve hand-installed .debs into manual_deb_sync
Hand-installed `.deb` packages become their own job with their own
`sync_jobs` flag, validation, review and JobResult, mirroring the shape
#251 and #252 will follow for sideloaded snaps and manual flatpaks.
`manual_installs_sync` keeps the unowned-path half alone.

The half both jobs need — the item shape, the plan pipeline, the shared
install-snippet registry with its push and consent question, the review
grouping and the replay — moves to `packages/unreproducible.py` as
`UnreproducibleSyncJob`, so neither job owns the other's machinery.

`item_id` is byte-identical across the split, so the shared snippet
registry needs no migration. Machine-specific marks do: they live in a
per-manager decision file, so `DecisionFile` gains `AdoptedMarks` and the
new job reads its own slice of `manual.decisions.yaml` as well as its own
file — otherwise every "never install this" answer the user already gave
about a `.deb` would be silently orphaned (D-08).

Refs #253
