+++ src/pcswitcher/config_sync.py
@@ -30,7 +30,7 @@ if TYPE_CHECKING:
 #
 # config_sync carries exactly ONE file, config.yaml (D-23): it is the single required
 # config a first sync needs. The shared install-snippet registry is NOT carried here —
-# it travels by `manual_installs_sync`'s own post-review `send_file` push, because
+# it travels by an unreproducible job's own post-review `send_file` push, because
 # config_sync runs before any review (sync step 9) and so cannot carry a snippet the
 # user authored during that review.
 CONFIG_REMOTE_DIR: str = "~/.config/pc-switcher"
@@ -308,7 +308,7 @@ async def sync_config_to_target(  # noqa: PLR0913 - config state + flags; flags 
     3. Target config matches: Skip silently
 
     Carries exactly one file, the caller-supplied `source_config_path` (config.yaml). The
-    install-snippet registry is NOT transferred here — `manual_installs_sync` pushes it
+    install-snippet registry is NOT transferred here — an unreproducible job pushes it
     itself after its review (D-23).
 
     Args:
