+++ src/pcswitcher/default-config.yaml
@@ -47,12 +47,13 @@ sync_jobs:
   # then their data lands on top of them (installing software writes default config/data
   # files; running folder_sync afterward lets the user's synced versions land on top of
   # those defaults). The orchestrator resolves sync_jobs in the order keys appear in this file, so
-  # apt_sync, snap_sync, flatpak_sync and manual_installs_sync are listed above
-  # folder_sync here — reordering these entries changes execution order.
+  # apt_sync, snap_sync, flatpak_sync, manual_deb_sync and manual_installs_sync are
+  # listed above folder_sync here — reordering these entries changes execution order.
   apt_sync: false             # Install missing apt packages on target, after a batched review (destructive; opt-in)
   snap_sync: false            # Converge snap revisions/channels on target, after a batched review (destructive; opt-in)
   flatpak_sync: false         # Converge flatpak refs/remotes on target, after a batched review (destructive; opt-in)
-  manual_installs_sync: false # Detect and reproduce installs no package manager can, via install snippets (destructive; opt-in)
+  manual_deb_sync: false      # Reproduce hand-installed .deb packages on target, via install snippets (destructive; opt-in)
+  manual_installs_sync: false # Reproduce unowned installs under /usr/local and /opt, via install snippets (destructive; opt-in)
 
   folder_sync: true          # Sync /home and /root directories (Phase 1)
   vscode_state_sync: true    # Sync VS Code editor state (your user only), except its per-machine secret store
