+++ src/pcswitcher/jobs/__init__.py
@@ -10,6 +10,7 @@ from .dummy_fail import DummyFailJob
 from .dummy_success import DummySuccessJob
 from .folder_sync import FolderEntry, FolderSyncJob
 from .install_on_target import InstallOnTargetJob
+from .manual_deb_sync import ManualDebSyncJob
 from .manual_installs_sync import ManualInstallsSyncJob
 from .vscode_state_sync import VscodeStateSyncJob
 
@@ -24,6 +25,7 @@ __all__ = [
     "InstallOnTargetJob",
     "Job",
     "JobContext",
+    "ManualDebSyncJob",
     "ManualInstallsSyncJob",
     "SyncJob",
     "SystemJob",
