+++ src/pcswitcher/schemas/config-schema.yaml
@@ -77,10 +77,14 @@ properties:
         type: boolean
         default: false
         description: "Converge installed flatpak refs and remotes, after a batched review (destructive; opt-in)"
+      manual_deb_sync:
+        type: boolean
+        default: false
+        description: "Detect and reproduce hand-installed .deb packages, after a batched review (destructive; opt-in)"
       manual_installs_sync:
         type: boolean
         default: false
-        description: "Detect and reproduce installs no package manager can (bare .debs, unowned /usr/local, /opt), after a batched review (destructive; opt-in)"
+        description: "Detect and reproduce unowned installs under /usr/local and /opt, after a batched review (destructive; opt-in)"
       # Jobs implemented in Phase 1:
       folder_sync:
         type: boolean
@@ -231,7 +235,8 @@ properties:
   # VscodeStateSyncJob has no configuration: the module hardcodes the editor list, DB
   # layout, and the preserved-key namespace (secret://). Enable/disable via sync_jobs.
 
-  # The apt_sync/snap_sync/flatpak_sync/manual_installs_sync jobs have no config section:
+  # The apt_sync/snap_sync/flatpak_sync/manual_deb_sync/manual_installs_sync jobs have no
+  # config section:
   # they earn one only when they gain a real key (D-32). A config omitting them still
   # validates, and config.py's get_job_config returns {} for an absent section, which each
   # job's empty CONFIG_SCHEMA accepts. Enable/disable via sync_jobs only.
